% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolliter.R
\name{rollmeanp}
\alias{rollmeanp}
\title{Rolling Mean}
\usage{
rollmeanp(x, w = 5, front = TRUE)
}
\arguments{
\item{x}{numeric; data vector}

\item{w}{integer; width of the rolling window}

\item{front}{logical; when \code{w} is odd, should the new values be biased to the front?}
}
\description{
Rolling mean that includes partial results from the beginning and end
}
\details{
Wrapper/extension for \code{TTR::runMean()}
}
\examples{
xx <- c(3, 8, 5, 9, 6, 3, 8, 2, 5, 1)
plot(xx)
lines(rollmeanp(xx, w=4, front=TRUE), col="blue")
lines(rollmeanp(xx, w=4, front=FALSE), col="red")
 
}
