% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollrms.R
\name{rollrms}
\alias{rollrms}
\title{Rolling RMS}
\usage{
rollrms(x, w, na.pad = TRUE, partial = FALSE)
}
\arguments{
\item{x}{numeric vector; data}

\item{w}{integer; width of the rolling RMS window}

\item{na.pad}{logical; should the ends be padded with \code{NA}?}

\item{partial}{logical; should partial results be returned?}
}
\description{
Wrapper for \code{TTR::runMean()} offering a quick rolling RMS
}
\examples{
set.seed(1)
data(xyz2)
r <- xyz2
r.rms <- rollrms(r, w=25, partial=TRUE)
r.rms.m <- mean(r.rms, na.rm=TRUE)
plot(r, main="Rolling RMS")
lines(r.rms, col="red", lwd=1.2)
abline(h=r.rms.m, col="blue")
mtext(round(r.rms.m, 3),
  4, at=r.rms.m, las=1, cex=0.6, line=0.1)
}
