% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollwinrms.R
\name{rollwinrms}
\alias{rollwinrms}
\title{Rolling RMS}
\usage{
rollwinrms(x, w, ...)
}
\arguments{
\item{x}{numeric vector; data}

\item{w}{integer or numeric vector; width or shape of the rolling RMS window}

\item{na.pad}{logical; should the ends be padded with \code{NA}?}

\item{B}{prime; smoothness as used by the FFT algorithm}
}
\description{
Wrapper for \code{rollconv()} offering a quick rolling windowed RMS
}
\examples{
w <- 50
weights.para <- winweights(w, type="parabolic")
weights.hann <- winweights(w, type="hann")

plot(weights.hann, type="l", ann=FALSE)
lines(weights.para)

data(xyz2)
yy <- xyz2

plot(yy, col="grey", type="l")
lines(rollwinrms(yy, weights.hann, partial=TRUE), col="blue", lwd=2)
lines(rollwinrms(yy, weights.para, partial=TRUE), col="green", lwd=2)
lines(rollwinrms(yy, w, partial=TRUE), col="red", lwd=2)
}
