% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_doc.R
\docType{data}
\name{stereo}
\alias{stereo}
\title{A short stereo audio sample.}
\format{A data frame with 22050 rows and 2 columns:
\describe{
  \item{l}{left channel, values in signed 16 bit integers}
  \item{r}{right channel, values in signed 16 bit integers}
}}
\source{
A Tribe Calld Quest – The Killing Season
}
\usage{
stereo
}
\description{
The first 22050 samples (0.5 seconds) of a stereo
audio file
}
\keyword{datasets}
