% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezwin.R
\name{trapezwin}
\alias{trapezwin}
\title{Trapepezoid windows}
\usage{
trapezwin(l = 10, prop = 0.5, l.slopes)
}
\arguments{
\item{l}{integer; length of the window}

\item{prop}{numeric; proportion of the length covered by slopes}

\item{l.slopes}{integer; length of the slopes. Overrides \code{prop}}

\item{ground}{logical; should the legs of the trapezoid be 'grounded'}
}
\description{
Create isosceles (symmetric) trapezoid windows
}
\examples{
x <- 100
plot(0, type="n", xlim=c(0, x), ylim=c(0, 2/x))
for (i in (1:20)/20){
    lines(trapezwin(x, prop=i))
}

x <- 10000
plot(0, type="n", xlim=c(0, x), ylim=c(0, 2/x))
for (i in ceiling(seq(1, x \%/\% 2, length.out=20))){
    lines(trapezwin(x, l.slopes=i))
}

}
