% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winweights.R
\name{winweights}
\alias{winweights}
\title{Discrete windows}
\usage{
winweights(width = 11, type = "epanechnikov", step.adj = TRUE,
  lev.adj = "sum", a = 3)
}
\description{
Generate coefficients for various popular window functions
}
\examples{
# Time and frequency plots of each window 
name <- c("square", "triangular",
          "epanechnikov", "biweight", "triweight",
          "tricube", "cosine-smooth", "optcosine",
          "hann", "hamming", "blackman", "nuttall", "blackman-nuttall",
          "blackman-harris", "flat-top", "kaiser-bessel",
          "lanczos", "sinc")

w <- 99
par(mfcol=c(6, 3), mar=c(1.5, 1.5, 0.5, 0.5), mgp=c(0, 0.6, 0))
for (i in name) {
    plot(winweights(w, type=i, a=3), type="l", xlab="", ylab="", ylim=c(-0.02, 0.06))
    grid(col="#00000022", lty=1)
    legend("topright", legend=i, bty="n", text.col="blue", adj=c(0.2, 0))
}

set.seed(1)
w <- 99
x <- rnorm(2e4)
i <- 1
par(mfcol=c(6, 3), mar=c(1.5, 1.5, 0.5, 0.5), mgp=c(0, 0.6, 0))
for (i in 1:length(name)) {
    win <- winweights(w, type=name[i], step.adj=TRUE, lev.adj="mean", a=3)
    rol <- rollconv(x, win, partial=FALSE, scale.window=FALSE)
    spectrum(na.omit(rol), main="", xlab="", ylab="", sub="", lwd=0.2, ylim=c(1e-12, 1e5))
    grid(col="#00000022", lty=1)
    legend("topright", legend=name[i], bty="n", text.col="blue", adj=c(0.2, 0))
}

}
