% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{dema}
\alias{dema}
\title{Double Exponential Moving Average}
\usage{
dema(x, a = 2, direction = c("ff", "fb", "bf", "av"))
}
\arguments{
\item{x}{a numeric vector}

\item{a}{numeric; smoothing factor}

\item{direction}{character; should the two moving averages be run
\itemize{
  \item{\code{ff}: forwards both times}
  \item{\code{fb}: forwards then backwards}
  \item{\code{bf}: backwards then forwards}
  \item{\code{av}: both \code{fb} and \code{bf}, returning the average}
}}
}
\description{
\code{dema} performs exponential smoothing twice on the same vector. \cr
This is not the same as the DEMA typically used in technical trading 
(eg. \code{\link{TTR::DEMA}}).
}
\details{
The moving averages are done as in \code{ema}, except that 
\code{a} is square-rooted, to make the outputs more comparable.
}
\examples{
s <- c(0, 0, 1, NA, 2, 0, 1, 1, NA, NA, 2, NA, 3, 1, 2, 1, 0, 1, NA, 0, 0, 0)
plot(s, type="b", col="#00000088")
lines(dema(s, 2, dir="ff"), col=2)
lines(dema(s, 2, dir="fb"), col=3)
lines(dema(s, 2, dir="bf"), col=4)
lines(dema(s, 2, dir="av"), col=5)
}
\seealso{
\code{\link{ema}}, \code{\link{iema}}
}
