% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{ema}
\alias{ema}
\title{Run Exponential Moving Average over vector}
\usage{
ema(x, a = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{a}{smoothing factor. Equvalent to \eqn{1 / \alpha}}
}
\description{
\code{ema} returns a vector of exponentially smoothed values
}
\examples{
x <- c(rep(0, 4), 1, rep(0, 59))
a <- 1:10*5
col <- rainbow(length(a), start=0.15)

plot(x, type="p", pch=16, cex=0.4, ylim=c(0, 0.21), xaxs="i",
  main="Exponential Moving Average", xlab="Time", ylab="Magnitude")
for (i in seq_along(a)) {
    lines(ema(x, a[i]), col=col[i])
}

mtext("Single impulse of magnitude 1 occuring at time = 5", line=0.2)
L <- lapply(seq_along(a), function(x) bquote(alpha ==  1 / .(a[x])))
legend("topright", legend=as.expression(L),
  bty="n", col=col, lwd=1.2, cex=0.8, y.intersp=1.2)
}
